+++ pkg/operator/controllers/httperrors.go
@@ -1,4 +1,4 @@
-package http
+package controllers
 
 import (
 	"net/http"
